requirements.txt
# Core RL packages
numpy==1.23.5
gym==0.21.0
stable-baselines3==1.7.0

# Plotting and logging
matplotlib
pandas
PyYAML
tensorboard




